chore: rename authority
