+++ src/test/java/de/caritas/cob/uploadservice/api/authorization/RoleAuthorizationAuthorityMapperTest.java
@@ -8,7 +8,6 @@ import java.security.Principal;
 import java.util.HashSet;
 import java.util.Set;
 import java.util.stream.Collectors;
-import org.hamcrest.MatcherAssert;
 import org.junit.Test;
 import org.junit.runner.RunWith;
 import org.keycloak.adapters.RefreshableKeycloakSecurityContext;
@@ -22,7 +21,7 @@ import org.springframework.security.core.Authentication;
 import org.springframework.security.core.authority.SimpleGrantedAuthority;
 
 @RunWith(MockitoJUnitRunner.class)
-public class RoleAuthorizationAuthoritiesMapperTest {
+public class RoleAuthorizationAuthorityMapperTest {
 
   private final KeycloakAuthenticationProvider provider = new KeycloakAuthenticationProvider();
   private final Set<String> roles =
@@ -50,7 +49,7 @@ public class RoleAuthorizationAuthoritiesMapperTest {
 
     Set<SimpleGrantedAuthority> expectedGrantendAuthorities = new HashSet<>();
     roles.forEach(roleName -> {
-      expectedGrantendAuthorities.addAll(Authorities
+      expectedGrantendAuthorities.addAll(Authority
           .getAuthoritiesByUserRole(UserRole.getRoleByValue(roleName).get()).stream()
           .map(SimpleGrantedAuthority::new).collect(Collectors.toSet()));
     });
